+++ src/flask/sessions.py
@@ -223,6 +223,14 @@ class SessionInterface:
         the :data:`SESSION_COOKIE_SAMESITE` setting.
         """
         return app.config["SESSION_COOKIE_SAMESITE"]  # type: ignore[no-any-return]
+    
+    def get_cookie_partitioned(self, app: Flask) -> bool:
+        """Returns True if the cookie should be partitioned. This currently
+        just returns the value of the ``SESSION_COOKIE_PARTITIONED`` setting.
+
+        .. versionadded:: 3.1.0
+        """
+        return app.config["SESSION_COOKIE_PARTITIONED"]  # type: ignore[no-any-return]
 
     def get_expiration_time(self, app: Flask, session: SessionMixin) -> datetime | None:
         """A helper method that returns an expiration date for the session